Added user control on delete

+++ src/AppBundle/Controller/TwitController.php
@@ -220,6 +220,9 @@ class TwitController extends Controller
                 throw $this->createNotFoundException('Unable to find Twit entity.');
             }
 
+            if ($entity.srcUsrId !== $this->get('security.context')->getToken()->getUser() ) {
+                throw $this->createNotFoundException('You are not allowed to delete someone else twits.');
+            }
             $em->remove($entity);
             $em->flush();
